[MT] Using cond vars. Fixed problem.
Used notify_all instead of notify_one.

The problem previously was happening at the end, where the notify
might pick a thread, but at the same time, the thread was being
probably joined.

Notify all solves it so that it isn't a problem anymore.

Tested with 1000x producers, 1000x consumers, each playing with
500 items.